Make the server use the registry_deps option when calling the `register` function (#67)
* Make server use registry dependencies

The register method had an option called registry_deps, which could be used to provide a set of repositories to be used for resolving dependencies. The server had no way of using this option, so I added an option to the config and gave the server the ability. By default General is already there in the config.

* Make registry_deps optional in config