Handle case where no tags exists in approval process
Hi,

I've been running into an issue recently where the approval process on a custom version of Registrator will error when there are no existing tags for the repository from which the original registration event originated. The code works just fine if there is an original tag.

An alternative fix is to perhaps call `tags` with the `handle_error=false` option, though I haven't tried that locally. Happy to modify this PR to use that option if it's preferable, let me know!