add description field (#349)
Over in https://github.com/JuliaRegistries/General/issues/43776 it was suggested that it might be nice to have the repo description field in the PR comment. We could then grab it from the PR comment and send it downstream as well, e.g. to the Slack channel `#new-packages-feed`, along with the release notes.

From the WebUI or CommentBot we seem to have a `repo` object with this field, so we can get the description without an additional API call. However, I'm not sure if it is always populated or not; if it is not populated, then we'd need to make an API call to fill out the fields (`repo` in the comment bot to use GitHub.jl, or `get_repo` in the WebUI to use GitForge.jl).