Push to fork of registry
We would like to use a fork of the registry repo to make registration pull requests. This would allow us to make registration requests without requiring that Registrator have write access to the Registry repo.

Added a new configuration parameter to configure the Registry fork's URL. If this is unspecified it defaults to the clone URL.